refactor: move ColorPicker to a file
