+++ src/components/StringToQRCode.tsx
@@ -1,12 +1,11 @@
 import React, { useState, useCallback, useRef } from "react";
 import { nanoid } from "nanoid";
 import QRCode from "react-qr-code";
-import Tippy from "@tippyjs/react";
 import { toPng } from "html-to-image";
 // import { addNote, notes } from "../store";
 // import { useStore } from "@nanostores/react";
-import { ChromePicker } from "@hello-pangea/color-picker";
 import { VscLoading as LoadingIcon } from "react-icons/vsc";
+import { ColorPicker } from "./ColorPicker";
 
 const StringToQRCode = () => {
   const qrRef = useRef<HTMLDivElement>(null);
@@ -116,43 +115,3 @@ const StringToQRCode = () => {
 };
 
 export default StringToQRCode;
-
-interface IColorPickerProps {
-  hoverContent: string;
-  color: string;
-  setColor: (color: string) => any;
-}
-const ColorPicker: React.FC<IColorPickerProps> = ({
-  color,
-  setColor,
-  hoverContent,
-}) => {
-  return (
-    <div className="relative h-16 w-16">
-      <Tippy content={hoverContent} hideOnClick={true}>
-        <div>
-          <Tippy
-            theme="transparent"
-            interactive={true}
-            trigger="click"
-            arrow={false}
-            content={
-              <div className="">
-                <ChromePicker
-                  disableAlpha={true}
-                  color={color}
-                  onChange={(v) => setColor(v.hex)}
-                />
-              </div>
-            }
-          >
-            <div
-              className="absolute h-16 w-16 rounded-md border cursor-pointer"
-              style={{ backgroundColor: color }}
-            ></div>
-          </Tippy>
-        </div>
-      </Tippy>
-    </div>
-  );
-};
