+++ src/pages/index.tsx
@@ -1,6 +1,5 @@
 import StringToQRCode from "@/components/StringToQRCode";
 import Head from "next/head";
-import Image from "next/image";
 import Link from "next/link";
 
 export default function Home() {
